refactor: replace AsyncWrapper with AsyncRow in tokens and apps components
